requirements.txt
python-dotenv
openai
gradio
Pillow
anthropic